+++ packages/cli/main.js
@@ -2,6 +2,8 @@
 
 const { commands } = require('./dist/index');
 
+const issuerTypeGenerators = commands.data.typeGenerators;
+
 const options = {
   input: {
     alias: 'i',
@@ -23,15 +25,29 @@ const options = {
     alias: 'hd',
     description: 'HD Path to derive key',
   },
+  seed: {
+    alias: 's',
+    description: 'Seed to generate key from',
+  },
   type: {
     alias: 't',
     description: 'Type of key to derive',
   },
   format: {
     alias: 'f',
-    choices: ['vc', 'vc-jwt'],
+    choices: ['vc', 'vc-jwt', 'vp', 'vp-jwt'],
     description: 'Output format',
     default: 'vc',
+  },
+  domain: {
+    alias: 'd',
+    type: 'string',
+    description: 'Domain of the verifier',
+  },
+  challenge: {
+    alias: 'c',
+    type: 'string',
+    description: 'Challenge from the verifier',
   },
   endpoint: {
     alias: 'e',
@@ -43,135 +59,52 @@ const options = {
     stype: 'string',
     description: 'Authorization token to use',
   },
+  // These options to be used for "Certified" type generators
+  issuerType: {
+    alias: 'it',
+    choices: Object.keys(issuerTypeGenerators),
+    description: 'Type of issuer to create from',
+  },
+  issuerSeed: {
+    alias: 'is',
+    type: 'number',
+    description:
+      'Seed for deriving random values. This value is low entropy and for testing purposes only.',
+  },
+  subjectSeed: {
+    alias: 'ss',
+    type: 'number',
+    description:
+      'Seed for deriving random values. This value is low entropy and for testing purposes only.',
+  },
+  // These options to be used for Presentations
+  holderType: {
+    alias: 'ht',
+    choices: Object.keys(issuerTypeGenerators),
+    description: 'Type of holder to present from',
+  },
+  holderSeed: {
+    alias: 'hs',
+    type: 'number',
+    description:
+      'Seed for deriving random values. This value is low entropy and for testing purposes only.',
+  },
+  credentialsDirectory: {
+    alias: 'vcd',
+    type: 'string',
+    description: 'The directory containing credentials to present.',
+  },
 };
 
-require('yargs')
+const cli = require('yargs')
   .scriptName('transmute')
-  .usage('$0 <cmd> [args]')
-  // key
-  .command({
-    command: 'key [operationType]',
-    describe: 'Manage key material',
-    handler: argv => {
-      if (argv.operationType === 'generate') {
-        return commands.key.generate.generateKeyCommand[3](argv);
-      }
-    },
-    builder: {
-      operationType: {
-        demand: true,
-        choices: ['generate'],
-      },
-      ...commands.key.generate.generateKeyCommand[2],
-    },
-  })
+  .usage('$0 <cmd> [args]');
 
-  // credential
-  .command({
-    command: 'credential [operationType]',
-    describe: 'See https://www.w3.org/TR/vc-data-model/#credentials',
-    handler: argv => {
-      const dispatchHandler = {
-        create: commands.credential.createCredentialHandler,
-        verify: commands.credential.verifyCredentialHandler,
-        setStatusListIndex: commands.credential.setStatusListIndexHandler,
-        isStatusListIndexSet: commands.credential.isStatusListIndexSetHandler,
-      };
-      if (dispatchHandler[argv.operationType]) {
-        return dispatchHandler[argv.operationType](argv);
-      } else {
-        throw new Error('Unsupported operationType');
-      }
-    },
-    builder: {
-      operationType: {
-        demand: true,
-        choices: [
-          'create',
-          'verify',
-          'setStatusListIndex',
-          'isStatusListIndexSet',
-        ],
-      },
-      ...options,
-    },
-  })
+commands.key.registerCommands(cli);
+commands.credential.registerCommands(cli);
+commands.presentation.registerCommands(cli);
+commands.data.registerCommands(cli);
+commands.neo.registerCommands(cli);
+commands.google.registerCommands(cli);
 
-  // presentation
-  .command({
-    command: 'presentation [operationType]',
-    describe: 'See https://www.w3.org/TR/vc-data-model/#presentations',
-    handler: argv => {
-      if (argv.operationType === 'create') {
-        return commands.presentation.createPresentationCommand[3](argv);
-      }
-      if (argv.operationType === 'verify') {
-        return commands.presentation.verifyPresentationCommand[3](argv);
-      }
-    },
-    builder: {
-      operationType: {
-        demand: true,
-        choices: ['create', 'verify'],
-      },
-      ...commands.presentation.createPresentationCommand[2],
-      ...commands.presentation.verifyPresentationCommand[2],
-    },
-  })
-
-  // data
-  .command({
-    command: 'data [operationType]',
-    describe: 'See https://www.w3.org/TR/vc-data-model/#lifecycle-details',
-    handler: argv => {
-      if (argv.operationType === 'create') {
-        return commands.data.createDataCommand[3](argv);
-      }
-    },
-    builder: {
-      operationType: {
-        demand: true,
-        choices: ['create'],
-      },
-      ...commands.data.createDataCommand[2],
-    },
-  })
-
-  // neo
-  .command({
-    command: 'neo [operationType]',
-    describe: 'See https://neo4j.com/',
-    handler: argv => {
-      if (argv.operationType === 'workflow') {
-        return commands.neo.importWorkflowInstanceComand[3](argv);
-      }
-    },
-    builder: {
-      operationType: {
-        demand: true,
-        choices: ['workflow'],
-      },
-      ...commands.neo.importWorkflowInstanceComand[2],
-    },
-  })
-
-  // google
-  .command({
-    command: 'google [operationType]',
-    describe:
-      'A wrapper around a few google cloud libraries that help manage verifiable data.',
-    handler: argv => {
-      if (argv.operationType === 'create') {
-        return commands.google.googleCommands[3](argv);
-      }
-    },
-    builder: {
-      operationType: {
-        demand: true,
-        choices: ['create'],
-      },
-      ...commands.google.googleCommands[2],
-    },
-  })
-
-  .help().argv;
+cli.help().argv;
