+++ packages/cli/src/commands/credential/index.ts
@@ -1,3 +1,4 @@
 export * from './createCredential';
 export * from './verifyCredential';
 export * from './statusList';
+export * from './registerCommands';
