+++ packages/cli/src/commands/data/generate/generateOrganization.ts
@@ -1,5 +1,5 @@
 import faker from 'faker';
-import { generateKey } from '../../key/generateKey';
+import { generateKey } from '../../key/generate';
 import { sha256 } from '../../../util';
 
 export const generateOrganization = async (argv: { seed: any }) => {
