+++ packages/cli/src/commands/data/generate/workflows/scenarios/e-commerce/e-commerce.scenario.bpmn
@@ -14,7 +14,7 @@
 <bpmn:process id="Process_0" isExecutable="true">
 
 <bpmn:startEvent id="StartEvent_0" name="Start">
-<bpmn:outgoing>geid_195_0</bpmn:outgoing>
+<bpmn:outgoing>geid_218_0</bpmn:outgoing>
 </bpmn:startEvent>
           
 
@@ -71,9 +71,9 @@
         
   </script>
 
-<bpmn:incoming>geid_195_0</bpmn:incoming>
+<bpmn:incoming>geid_218_0</bpmn:incoming>
 
-<bpmn:outgoing>geid_195_1</bpmn:outgoing>
+<bpmn:outgoing>geid_218_1</bpmn:outgoing>
 </bpmn:scriptTask>
         
 
@@ -130,9 +130,9 @@
         
   </script>
 
-<bpmn:incoming>geid_195_1</bpmn:incoming>
+<bpmn:incoming>geid_218_1</bpmn:incoming>
 
-<bpmn:outgoing>geid_195_2</bpmn:outgoing>
+<bpmn:outgoing>geid_218_2</bpmn:outgoing>
 </bpmn:scriptTask>
         
 
@@ -182,9 +182,9 @@
         
   </script>
 
-<bpmn:incoming>geid_195_2</bpmn:incoming>
+<bpmn:incoming>geid_218_2</bpmn:incoming>
 
-<bpmn:outgoing>geid_195_3</bpmn:outgoing>
+<bpmn:outgoing>geid_218_3</bpmn:outgoing>
 </bpmn:scriptTask>
         
 
@@ -253,9 +253,9 @@
         
   </script>
 
-<bpmn:incoming>geid_195_3</bpmn:incoming>
+<bpmn:incoming>geid_218_3</bpmn:incoming>
 
-<bpmn:outgoing>geid_195_4</bpmn:outgoing>
+<bpmn:outgoing>geid_218_4</bpmn:outgoing>
 </bpmn:scriptTask>
         
 
@@ -305,9 +305,9 @@
         
   </script>
 
-<bpmn:incoming>geid_195_4</bpmn:incoming>
+<bpmn:incoming>geid_218_4</bpmn:incoming>
 
-<bpmn:outgoing>geid_195_5</bpmn:outgoing>
+<bpmn:outgoing>geid_218_5</bpmn:outgoing>
 </bpmn:scriptTask>
         
 
@@ -360,21 +360,21 @@
         
   </script>
 
-<bpmn:incoming>geid_195_5</bpmn:incoming>
+<bpmn:incoming>geid_218_5</bpmn:incoming>
 
-<bpmn:outgoing>geid_195_6</bpmn:outgoing>
+<bpmn:outgoing>geid_218_6</bpmn:outgoing>
 </bpmn:scriptTask>
         
 
 
 <bpmn:endEvent id="EndEvent_0" name="End">
-<bpmn:incoming>geid_195_6</bpmn:incoming>
+<bpmn:incoming>geid_218_6</bpmn:incoming>
 </bpmn:endEvent>
           
 
 
 <bpmn:sequenceFlow 
-id="geid_195_0" 
+id="geid_218_0" 
 name="Import Started"
 sourceRef="StartEvent_0" 
 targetRef="Task_0">
@@ -384,7 +384,7 @@ targetRef="Task_0">
 
 
 <bpmn:sequenceFlow 
-id="geid_195_1" 
+id="geid_218_1" 
 name="Fill Order"
 sourceRef="Task_0" 
 targetRef="Task_1">
@@ -394,7 +394,7 @@ targetRef="Task_1">
 
 
 <bpmn:sequenceFlow 
-id="geid_195_2" 
+id="geid_218_2" 
 name="Ocean Transport"
 sourceRef="Task_1" 
 targetRef="Task_2">
@@ -404,7 +404,7 @@ targetRef="Task_2">
 
 
 <bpmn:sequenceFlow 
-id="geid_195_3" 
+id="geid_218_3" 
 name="Air Transport"
 sourceRef="Task_2" 
 targetRef="Task_3">
@@ -414,7 +414,7 @@ targetRef="Task_3">
 
 
 <bpmn:sequenceFlow 
-id="geid_195_4" 
+id="geid_218_4" 
 name="Process US Customs Entry"
 sourceRef="Task_3" 
 targetRef="Task_4">
@@ -424,7 +424,7 @@ targetRef="Task_4">
 
 
 <bpmn:sequenceFlow 
-id="geid_195_5" 
+id="geid_218_5" 
 name="Ship to Importer Of Record"
 sourceRef="Task_4" 
 targetRef="Task_5">
@@ -434,7 +434,7 @@ targetRef="Task_5">
 
 
 <bpmn:sequenceFlow 
-id="geid_195_6" 
+id="geid_218_6" 
 name="Import Completed"
 sourceRef="Task_5" 
 targetRef="EndEvent_0">
@@ -545,43 +545,43 @@ targetRef="EndEvent_0">
           
 
 
-    <bpmndi:BPMNEdge id="geid_195_0_di" bpmnElement="geid_195_0">
+    <bpmndi:BPMNEdge id="geid_218_0_di" bpmnElement="geid_218_0">
       
     </bpmndi:BPMNEdge>
   
 
 
-    <bpmndi:BPMNEdge id="geid_195_1_di" bpmnElement="geid_195_1">
+    <bpmndi:BPMNEdge id="geid_218_1_di" bpmnElement="geid_218_1">
       
     </bpmndi:BPMNEdge>
   
 
 
-    <bpmndi:BPMNEdge id="geid_195_2_di" bpmnElement="geid_195_2">
+    <bpmndi:BPMNEdge id="geid_218_2_di" bpmnElement="geid_218_2">
       
     </bpmndi:BPMNEdge>
   
 
 
-    <bpmndi:BPMNEdge id="geid_195_3_di" bpmnElement="geid_195_3">
+    <bpmndi:BPMNEdge id="geid_218_3_di" bpmnElement="geid_218_3">
       
     </bpmndi:BPMNEdge>
   
 
 
-    <bpmndi:BPMNEdge id="geid_195_4_di" bpmnElement="geid_195_4">
+    <bpmndi:BPMNEdge id="geid_218_4_di" bpmnElement="geid_218_4">
       
     </bpmndi:BPMNEdge>
   
 
 
-    <bpmndi:BPMNEdge id="geid_195_5_di" bpmnElement="geid_195_5">
+    <bpmndi:BPMNEdge id="geid_218_5_di" bpmnElement="geid_218_5">
       
     </bpmndi:BPMNEdge>
   
 
 
-    <bpmndi:BPMNEdge id="geid_195_6_di" bpmnElement="geid_195_6">
+    <bpmndi:BPMNEdge id="geid_218_6_di" bpmnElement="geid_218_6">
       
     </bpmndi:BPMNEdge>
