+++ packages/cli/src/commands/data/index.ts
@@ -1,2 +1,3 @@
 export * from './createData';
 export * from './generate';
+export * from './registerCommands';
