+++ packages/cli/src/commands/google/index.ts
@@ -1,1 +1,2 @@
 export * from './googleCommands';
+export * from './registerCommands';
