+++ packages/cli/src/commands/neo/index.ts
@@ -1,1 +1,2 @@
 export * from './importWorkflowInstance';
+export * from './registerCommands';
