+++ packages/cli/src/commands/presentation/index.ts
@@ -1,2 +1,3 @@
 export * from './createPresentation';
 export * from './verifyPresentation';
+export * from './registerCommands';
